Removing default export, add `attw` assertion during build. (#312)
