+++ package.json
@@ -23,7 +23,7 @@
   "sideEffects": false,
   "repository": {
     "type": "git",
-    "url": "https://github.com/RobinTail/merge-sx.git"
+    "url": "git+https://github.com/RobinTail/merge-sx.git"
   },
   "homepage": "https://github.com/RobinTail/merge-sx",
   "author": {
@@ -32,7 +32,7 @@
   },
   "license": "MIT",
   "scripts": {
-    "build": "tsup",
+    "build": "tsup && attw --pack",
     "lint": "eslint src",
     "test": "tsc -p tsconfig.unit.json && jest src",
     "intTest": "tsc -p tsconfig.int.json && jest integration-test",
@@ -45,6 +45,7 @@
     "@mui/material": "^5.1.0"
   },
   "devDependencies": {
+    "@arethetypeswrong/cli": "^0.13.3",
     "@mui/system": "^5.15.0",
     "@swc-node/register": "^1.5.4",
     "@swc/core": "^1.3.14",
